Improve the parked mode_worldwarp_0 to the ROM's size and shape

+++ include/mode_worldwarp.h
@@ -180,7 +180,7 @@ void DisableBg(s32 bg);
 void func_08006120(s32 a, s32 b);
 void func_08065ACC(TextSlot* p, s32 n);
 u8 func_080DF750(void);
-void func_0810057C(s32 a, s32 b, void* c, s32 d, s32 e, void* f, s32 g, s32 h);
+void func_0810057C(s16 a, s16 b, void* c, s16 d, s16 e, void* f, s16 g, s16 h);
 void func_081004EC(s32 a, s32 b, s32 c, void* d, s32 e, s32 f);
 u8 func_08100608(s16 a);
 u8 func_0810063C(s16 a);
